python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = "trade-in"
authors = ["Omar LAHBIB <o.lahbib@autobiz.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-storage = "2.0.0"
autobiz-etl-tools = {git = "https://gitlab.aut.bz/data-platform/data-domain-applications/ddas/autobiz_etl_tools.git", rev = "v0.6.18dev19"}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-mock = "^3.8.2"
pylint = "^2.14.5"
black = "^22.6.0"
mypy = "^0.961"
pandas = "^1.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
